requirements.txt
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
cryptography==41.0.5
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1695534362840/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
et-xmlfile==1.1.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1700579780973/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1698244280508/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1701092366260/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1699283905679/work
jupyter_core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1698673744260/work
MarkupSafe==2.1.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mpmath==1.3.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
networkx==3.2.1
numpy==1.26.2
openpyxl==3.1.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.1.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pdfminer.six==20221105
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==10.1.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1699715570510/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1699963054032/work
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1695367212704/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pycparser==2.21
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pypdf==3.17.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1698062516368/work
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.12
tokenizers==0.15.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1695373625840/work
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1701095650114/work
transformers==4.35.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
tzdata==2023.3
urllib3==2.1.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1700607916581/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
